move children backendAccess from parent to children
