+++ client/src/parent/BackendAccess.ts
@@ -243,91 +243,3 @@ export function useChildrenDataForUser(): ChildrenLoadingData {
 
   return useChildrenData(family);
 }
-
-export function useChildChoreData(id: number): ChoreLoadingData {
-  const bearerToken = useRecoilValue(AppState.userBearerToken);
-  const { isLoading, error, data: chores } = useQuery("childChoreData", () =>
-    axios
-      .get<ChoreWithAssignments[]>(`${apiBaseUrl}/api/Chores/Assignments/Users/${id}`, {
-        headers: {
-          Authorization: `Bearer ${bearerToken}`,
-        },
-      })
-      .then((data) => data?.data)
-  );
-
-  return { isLoading: isLoading, error: error, chores: chores ?? [] };
-}
-
-export function useUpdateChoreState(): UseMutationResult<
-  AxiosResponse<Assignment>,
-  unknown,
-  Assignment,
-  unknown
-> {
-  const bearerToken = useRecoilValue(AppState.userBearerToken);
-  const queryClient = useQueryClient();
-  const mutation = useMutation(
-    (assignment: Assignment) =>
-      axios.put<Assignment>(
-        `${apiBaseUrl}/api/Assignments/${assignment.id}`,
-        assignment,
-        {
-          headers: {
-            Authorization: `Bearer ${bearerToken}`,
-          },
-        }
-      ),
-    {
-      onSuccess: () => {
-        queryClient.invalidateQueries("childChoreData");
-      },
-    }
-  );
-
-  return mutation;
-}
-
-export function useChildRewardData(id: number): RewardLoadingData {
-  const bearerToken = useRecoilValue(AppState.userBearerToken);
-  const { isLoading, error, data: rewards } = useQuery("childRewardData", () =>
-    axios
-      .get<RewardWithGrants[]>(`${apiBaseUrl}/api/Rewards/Users/${id}`, {
-        headers: {
-          Authorization: `Bearer ${bearerToken}`,
-        },
-      })
-      .then((data) => data?.data)
-  );
-
-  return { isLoading: isLoading, error: error, rewards: rewards ?? [] };
-}
-
-export function useUpdateRewardState(): UseMutationResult<
-  AxiosResponse<Grant>,
-  unknown,
-  Grant,
-  unknown
-> {
-  const bearerToken = useRecoilValue(AppState.userBearerToken);
-  const queryClient = useQueryClient();
-  const mutation = useMutation(
-    (grant: Grant) =>
-      axios.put<Grant>(
-        `${apiBaseUrl}/api/Grants/${grant.id}`,
-        grant,
-        {
-          headers: {
-            Authorization: `Bearer ${bearerToken}`,
-          },
-        }
-      ),
-    {
-      onSuccess: () => {
-        queryClient.invalidateQueries("childRewardData");
-      },
-    }
-  );
-
-  return mutation;
-}
